smart-home-backend/requirements.txt
Flask==3.1.0
Flask-JWT-Extended==4.7.1
paho-mqtt==2.1.0
pymongo==4.13.0